Update wrapper for PR issues
This fixes a few problems:
  * Adds a base `ReqonError` class
  * Changes the `TypeError` to be called `InvalidTypeError`
  * Adds a decorator function to do the type checking

I think the decorator function makes the rest of the code a bit nicer to
look at and reason about, but the type checker decorator function is bit
much to look at and reason about. Not sure if this is the way to go or
not.